+++ src/ground_element.h
@@ -20,12 +20,12 @@
 
 	struct VS_GroundOutput {
 		float4	position	: POSITION;
-		half4	color		: COLOR0;
-		half3	eyevec		: COLOR1;
-		half3	sundir		: TEXCOORD0;
-		half3	skydir		: TEXCOORD1;
-		half3	sunlight	: TEXCOORD2;
-		half3	skylight	: TEXCOORD3;
+		float4	color		: COLOR0;
+		float3	eyevec		: COLOR1;
+		float3	sundir		: TEXCOORD0;
+		float3	skydir		: TEXCOORD1;
+		float3	sunlight	: TEXCOORD2;
+		float3	skylight	: TEXCOORD3;
 		
 		float4	inscatter	: TEXCOORD4;
 		float4	extinction	: TEXCOORD5;
